+++ libdf3d/platform/desktop_common/FileDataSourceDesktop.h
@@ -15,10 +15,10 @@ public:
     bool valid() const override;
 
     size_t getRaw(void *buffer, size_t sizeInBytes) override;
-    int64_t getSizeInBytes() override;
+    size_t getSizeInBytes() override;
 
-    int64_t tell() override;
-    bool seek(int64_t offset, std::ios_base::seekdir origin) override;
+    size_t tell() override;
+    bool seek(size_t offset, std::ios_base::seekdir origin) override;
 };
 
 } }
